Feature #23589: Managed Library: Delete: Single and all library.  - Written onClick function on click of delete hyper link.  - Fetch parameter to redirect the link.  - Made ajax call for it.  - Written method in model to fetch the name of library.  - Passed name of library and id of it to success response of ajax.
 Managed Library: Preview
 - Craeted view for it.
 - Written action on controller to render view.

+++ components/filehandler.php
@@ -547,12 +547,16 @@ public static function deletecoursefile($file) {
 			return false;
 		}
 	} else {
-		$base = AppConstant::UPLOAD_DIRECTORY;
-		if (unlink($base."$file")) {
-			return true;
-		} else {
-			return false;
-		}
+
+        $base = AppConstant::UPLOAD_DIRECTORY;
+        if(file_exists($base)){
+            if (unlink($base."$file")) {
+                return true;
+            } else {
+                return false;
+            }
+        }
+
 	}
 }
 public static function deleteqimage($file) {
